fix: split settings_generator helpers, remove premature plugin modules
- Extract settings_helpers.rs (89L) from settings_generator.rs (429→353L)
- Remove plugin_cli, plugin_manifest, plugin_manager from lib.rs
  (not yet ported; were added prematurely by parallel agent)
- 246 tests pass, clippy clean
